Migrate from ES2018 to ES2017
The repository has been migrated from ES2018 to ES2017. The only change
was replacing the Object spread operator with `Object.assign`. This was
done primarily because the MetaMask extension uses a build tool
(`esprima`) that isn't compatible with this operator. The MetaMask
extension also supports certain browser versions (Chrome <v60) that
don't support this operator.