Rename exports to be more generic (#75)
* Rename exports to be more generic

* Update changelog

* Specify Ethereum EIP-1474 in a comment, mention Ethereum EIP-1193 renaming

This commit updates comments in the README file to make it clear that JSON-RPC errors and Ethereum EIP-1474 errors are under the 'rpcErrors' namespace, and adds that Ethereum EIP-1193 Provider errors are now under the 'providerErrors' namespace.

* Export providerErrors from index.ts file

The providerErrors was not exported from the index.ts file, so it was imported separately from the errors module. This commit includes the export statement for providerErrors in the index.ts file.

* Update code coverage thresholds

Adjusting the coverage thresholds to meet the new requirements.