feat: Preserve original messages during error serialization by default (#158)
This ensures that non-empty string `error.message` properties of
serialized errors are preserved by default, even if the serialized error
is not [a valid JSON-RPC
error](https://www.jsonrpc.org/specification#error_object). This
behavior can be overridden by setting `shouldPreserveMessage: false`.

In #61, our error serialization logic was considerably improved. One of
the behavioral changes made at the time was to always overwrite the
`message` property with that of the fallback error (practically always
the "internal JSON-RPC-error"), regardless of whether a non-empty string
message was present on the original error object. We have yet to ship
this everywhere in our stack, in part because such a change may be
breaking for our consumers. By reverting to the old behavior for the
`message` property only, we avoid these potential breakages and improve
the accessibility of potentially useful information to consumers (i.e.
directly in the error message as opposed to buried in
`error.data.cause.message`).